--- a/rten-gemm/Cargo.toml
+++ b/rten-gemm/Cargo.toml
@@ -20,6 +20,7 @@ rten-testing = { path = "../rten-testing" }
 
 [features]
 avx512 = ["rten-simd/avx512"]
+i8mm = []
 
 [lib]
 crate-type = ["lib"]