Add a new GEMM kernel for int8 matmul that uses [i8mm](https://developer.arm.com/documentation/ddi0602/2025-06/SIMD-FP-Instructions/UMMLA--vector---Unsigned-8-bit-integer-matrix-multiply-accumulate-to-32-bit-integer--vector--?lang=en) instructions.

The current implementation is very slightly slower on M3 Pro than the dotprod kernels. On an AWS Graviton 4 (c8g.xlarge), the i8mm kernel is about 22% faster for int8 matmuls, as tested on a modernbert-base model (with sequence_length 512). The marginal drop on M3 Pro might be due to the different kernel tile shape (8x8 instead of 16x4) rather than compute, but I need to verify.

Because the performance isn't always as good or better than dotprod, this is initially opt-in via an `i8mm` crate feature.

**Summary of changes:**

 - Add a new GEMM kernel that uses i8mm. Each kernel call updates an 8x8xi32 output tile from 8x8 int8 LHS and RHS inputs. The LHS is packed as a `(K/8, MR, 8)` panel. The RHS is packed as a `(K/8, NR, 8)` panel. The kernel loops over K tiles of size 8 and each iteration performs 16 2x8 @ 8x2 mini-matmuls to update an accumulator arranged as a 4x4 grid of 2x2 tiles. After the main accumulator loop, the 4x4 grid is shuffled to an 8x2 grid in row-major order for adding the zero point adjustment and writing back to memory.
 
Part of https://github.com/robertknight/rten/issues/685