@robertknight 2025-07-26
On an M3 Pro system:

```
$ set -x RAYON_NUM_THREADS 5  # Number of performance cores I have
$ cargo test -p rten-gemm -r -- --ignored bench_gemm_mix --nocapture

# main (MR=4, NR=16)
Testing kernel arm-int8-udot
m 512 n 512 k 512 iters 512. Duration 126.720ms (0.248ms/iter). GFLOPS 1084.6
m 1024 n 1024 k 1024 iters 64. Duration 100.660ms (1.573ms/iter). GFLOPS 1365.4
m 128 n 2048 k 512 iters 512. Duration 143.809ms (0.281ms/iter). GFLOPS 955.7
m 2048 n 128 k 512 iters 512. Duration 150.212ms (0.293ms/iter). GFLOPS 915.0

# This branch (MR=8, NR=8)
Testing kernel arm-int8-i8mm
m 512 n 512 k 512 iters 512. Duration 143.917ms (0.281ms/iter). GFLOPS 955.0
m 1024 n 1024 k 1024 iters 64. Duration 109.517ms (1.711ms/iter). GFLOPS 1255.0
m 128 n 2048 k 512 iters 512. Duration 185.054ms (0.361ms/iter). GFLOPS 742.7
m 2048 n 128 k 512 iters 512. Duration 163.434ms (0.319ms/iter). GFLOPS 840.9

# This branch, but with MR=4, NR=16
m 512 n 512 k 512 iters 512. Duration 135.528ms (0.265ms/iter). GFLOPS 1014.1
m 1024 n 1024 k 1024 iters 64. Duration 102.389ms (1.600ms/iter). GFLOPS 1342.3
m 128 n 2048 k 512 iters 512. Duration 150.798ms (0.295ms/iter). GFLOPS 911.4
m 2048 n 128 k 512 iters 512. Duration 156.169ms (0.305ms/iter). GFLOPS 880.1
```